chore: move form-informationmodel component
