+++ src/components/form-informationmodel/form-informationmodel.component.jsx
@@ -1,10 +1,10 @@
 import { connect } from 'react-redux';
 import { getFormSyncErrors, reduxForm } from 'redux-form';
 
-import { asyncValidateDatasetInvokePatch } from '../formsLib/asyncValidateDatasetInvokePatch';
+import { asyncValidateDatasetInvokePatch } from '../../pages/dataset-registration-page/formsLib/asyncValidateDatasetInvokePatch';
 
 import FormInformationModelPure from './form-informationmodel-pure.component';
-import { informationModelType } from '../../../schemaTypes';
+import { informationModelType } from '../../schemaTypes';
 import validate from './form-informationmodel-validations';
 
 const mapStateToProps = (state, { datasetItem }) => {
